requirements.txt
Pyyaml
setuptools
rich
xxhash
click
paramiko

